third_party/blink/web_tests/external/wpt/css/css-grid/masonry/tentative/track-sizing/auto-repeat/column-auto-repeat-009-expected.txt

This is a testharness.js-based test.
[FAIL] .masonry 1
  assert_equals: \n<div class="masonry" style="width: 200px; height: 100px;" data-expected-width="300" data-expected-height="200">\n    <div class="item" data-offset-x="200" data-offset-y="0" data-expected-width="100" data-expected-height="200"></div>\n  </div>\nwidth expected 100 but got 300
[FAIL] .masonry 2
  assert_equals: \n<div class="masonry" style="width: min-content; height: min-content;" data-expected-width="300" data-expected-height="200">\n    <div class="item" data-offset-x="200" data-offset-y="0" data-expected-width="100" data-expected-height="200"></div>\n  </div>\nwidth expected 100 but got 300
[FAIL] .masonry 3
  assert_equals: \n<div class="masonry" style="width: max-content; height: max-content;" data-expected-width="300" data-expected-height="200">\n    <div class="item" data-offset-x="200" data-offset-y="0" data-expected-width="100" data-expected-height="200"></div>\n  </div>\nwidth expected 100 but got 300
Harness: the test ran to completion.
